gestion-agricola/gestion-agricola-arduino/requirements.txt
dnspython==2.7.0
pymongo==4.10.1
pyserial==3.5
